syntax: make FuzzParsePrint take structured parameters
No more messing with bit masks.
The performance of the fuzzer seems largely the same,
though it does seem to discover interesting inputs much quicker.

That makes sense; it's now more aware of which knobs matter.

We also move the fuzzers to be in the same syntax test package,
so they can see all the other test data.