syntax: feed the fuzzer with lots of initial corpus
That is, with all the shell code from existing tests,
including "parser error", "print roundtrip", and "parse to AST".

This means we instantly reach over 2k interesting inputs,
and we are able to find new ones faster.

Note that we still leave the options fairly fixed,
with LangVariant being the only exception.
We expect that fuzzing can flip those pretty well.