syntax: add a LangVariant parameter to Quote
I had missed that $'' expansions are non-POSIX,
and only implemented by Bash and mksh.
So, in POSIX mode, we can't quote non-printable characters.

Moreover, fuzzing uncovered that mksh implements \x differently,
meaning that we require extra logic to follow its rules.
Keep all the fuzz crashers that we found in the process.

Since we've started having more edge cases that we can't quote,
start returning an error in the API, with a QuoteError type.
All it gives right now is a character position and a reason.

Finally, document what versions of Bash and mksh we develop with.
This matters, because some systems ship with very old versions,
which can implement slightly different quoting or escaping rules.

While at it, start using quicktest for the tests.