Revert "interp: start using pty.Start again"
This reverts commit 7b47fea60cdd767405e807ebdcccf3f0f4cb9dcc.

I was not able to reproduce any failures on Linux with that patch,
but Mac on CI failed immediately with:

	--- FAIL: TestRunnerTerminalExec (0.00s)
		--- FAIL: TestRunnerTerminalExec/Pseudo (0.61s)
			unix_test.go:129: EOF
	FAIL

For the time being, I'm giving up on trying to understand the lifetime
of these terminal files.

Keeping both files open until the command has finished running works
reliably on all plaforms. Just keep doing that for now.