syntax: extend FuzzQuote to check with our parser
That is, verify that our parser also parses the quoted string
as a single word in a simple command.

The updated fuzzer quickly found that we forgot about the empty string.
An empty string must be quoted, as otherwise it expands to zero fields.
The printf verification did not catch this.