interp: tweak tracing of assignments
Use the printer to print source-like array assignments.
The existing code was on the right track for non-arrays,
but we don't need to worry with WordParts at all.
Non-arrays get traced as the expanded value.

Add a test case exercising both kinds of assignments.
The previous code traced them in slightly incorrect ways.