README: point pre-commit hook towards github.com/scop/pre-commit-shfmt
The primary reason for this is to facilitate autoupdating. pre-commit's
`autoupdate` as well as other tooling capable of updating pre-commit
hook configs such as https://github.com/renovatebot/renovate work with
git tag `rev`s. I'm not aware of anything publically available at the
moment that would be able to update Go additional_dependencies strings.

The secondary reason is that the scop/pre-commit-shfmt hook is in
actual, active use by the repo maintainer and others. As such it is more
likely to be kept up to date than `local` hooks which detach from
everything the moment they're added to local configs.