interp: re-enable process substitution tests on Mac
We fixed a bug causing flakes with named pipes,
as our randomness didn't have enough bits
nor did it retry when files clashed.

I forget if issue 576 was truly a different bug or not.
I think it should also be fixed, so let's give it a try.
We can revert this commit and reopen the issue otherwise.

Fixes #576.