interp: check for empty paths in builtin test
test -d "" succeeds when it should fail. Checking all the flags which take
a path in gosh (ignoring G and O which are not yet implemented):

    for flag in b c d e f g h k L p r s S u w x; do
        test -$flag ""
        echo -n "$flag$? "
    done

Results in 'b1 c1 d0 e0 f1 g1 h1 k1 L1 p1 r0 s0 S1 u1 w1 x1'. Some are
correctly failing, but for the wrong reasons.

absPath is updated to check for empty paths and return them untouched.
changeDir was relying on the old behavior, so it now explicitly implements
the same behavior as bash dash gosh mksh and zsh, where a cd '' is
equivalent to a cd .

Shell behavior was surveyed with:

    try() {
        local cmd="$1" want="$2"
        for shell in bash csh dash gosh ksh mksh posh sh tcsh zsh; do
            local got
            got="$($shell -c "$cmd" 2>&1)"
            if [ "$got" != "$want" ]; then
                printf "~   FAIL: %s\n" "$shell: $cmd"
                printf "    want: %s\n" "$want"
                printf "     got: %s\n" "$got"
                echo
            fi
        done
    }

    try 'cd /; cd /tmp; cd ""; echo $? $PWD $OLDPWD' '0 /tmp /tmp'

    try 'for flag in b c d e f g h k L p r s S u w x; do
             test -$flag ""
             echo -n "$flag$? "
         done' \
         'b1 c1 d1 e1 f1 g1 h1 k1 L1 p1 r1 s1 S1 u1 w1 x1 '

Fixes #781