syntax: add support for heredocs directly inside backquotes
Our parser assumed that a heredoc must always end with a newline.
Unfortunately, the following is valid shell:

	`foo <<EOF
	body
	EOF`

Note the lack of a newline before the closing backquote.

The fix is relatively straightforward.
The two methods which tokenize heredoc bodies,
advanceLitHdoc and quotedHdocWord,
must learn to treat (r == '`' && p.backquoteEnd())
as an equivalent to the simpler case (r == '\n').

Note that we also make backquoteEnd more aggressive;
right now, it returns true even if we're in a nested quote state.
This is required because heredoc bodies use their own quote state,
and otherwise we wouldn't realise we're closing a backtick.

This seems like a good change, because backticks are special in shell.
They seem to tokenize at a much lower level,
which allows for bits of code like the one quoted above,
as well as:

	arg0 `# actually an inline comment without a newline!` \
		arg1

Fixes #729.