interp: add CallHandler
This handler is similar to ExecHandlerFunc, but has two major differences:

First, it runs for all simple commands, including function calls and builtins.

Second, it is not expected to execute the simple command, but instead to
allow running custom code which allows replacing the argument list.
Shell builtins touch on many internals of the Runner, after all.

Note that the handler also returns an error.
For the time being, this allows halting the runner, just like ExecHandler.
In the future, we could add more features via sentinel errors,
such as "skip this call entirely".

Fixes #705.