cmd/shfmt: change -ln default to "auto"
The man page describes the logic in detail.
In short, the default is no longer just "assume bash",
as it will try to detect the language from the filename and shebang,
and only fall back to "bash" if those fail.

Users who specify a language themselves, either via -ln or editorconfig,
should see no change in behavior whatsoever.
Those relying on the default should generally see an improvement,
as they might no longer have issues with directories mixing languages.

One potential regression is cases like:

	$ cat foo.sh
	foo=(bar baz) # actually bash

That is, bash scripts which use "sh" extensions or shebangs.
"sh" stands for POSIX shell, so the new logic will parse them as such,
whereas the old logic would parse them as bash.
Still, that seems like an easy fix on the part of the user:
they can either change the extension or shebang to be clear,
or they can explicitly set up the option to "bash".

Note that the added "auto" logic applies to all kinds of inputs:
explicit files, directories to be walked, and standard input.

Fixes #622.