interp: fix unset global inside function
In dc2e11e `unset` of a global var inside a function was broken, as
the special handling of `export` or `readonly` in function scope
introduced in that commit also incorrectly captured `unset`, calling a
parent.Set with the previous value.

Ensure we do not capture unset in the function scope global handling.

Unset of a global can fall-through this code. Globals are unset
correctly further down the Set function, so we do not need to invoke a
parent.Set at all.

Fixes #806