CI: start releasing a stable v3 docker tag
This way, "latest" continues as the latest development,
and users wanting to follow the latest v3.x.y version
can use the new "v3" tag.

Also add a docker deploy test branch.
Tested by pushing this commit to it earlier;
the v3-test docker tag was pushed as expected.

For #740.