syntax: strip escaped newlines from all non-quoted words
We did this in some cases involving simple literals,
but not in some others, such as two single-quoted literals joined with
an escaped newline.

The worst possible outcome is that formatting a program could break it.
For instance, the single statement below is equivalent to v=foobar:

	v='foo'\
	'bar'

however, we reformatted it to result in two commands due to indentation:

	v='foo'\
		'bar'

Instead, never keep escaped newlines in words if they aren't quoted.
They are still allowed when inside quotes, or when not part of a word.
Those are their main use cases, anyway, as we hadn't run into this
particularly buggy edge case yet.

Fixes #832.