cmd/shfmt: support long flags
That is, we now allow flags like --minify as well as -mn.
Note that we just use Go's standard flag package,
so that means we also support -minify and --mn.

For consistency, the docs now use --filename and --version,
which is backwards compatible as the single-dash versions still work.

Fixes #852.