interp: fix multi-line shell pattern match equality
commit be03afe7282aea8007fd3e961609a3b1f077db1d added the multi-line
flag to the regexp generated for `==` and `=!` testing. However, this
flag is confusing as it allows for a regular expression to match one
line out of a multi-line input.

    (m) multi-line mode: ^ and $ match begin/end line in addition to
    begin/end text (default false)

This means we are now matching against just one line in a multi-line
string, rather than the entire text. Instead, to match Bash's behavior,
we add the `s` flag to the regexp:

    (s) let . match \n (default false)

With this a `*` in a shell pattern will match `\n`.

Glob patterns should always match newline characters, so add the `s`
flag when the `.` metacharacter is used. This behavior worked by
accident with file globs because rather than matching with `.*` we match
with `[^/]*` which does match `\n`.