syntax: parse a redirect edge case without spaces
In the input

	>&2>/dev/null

we would incorrectly error on the `2`, as we would tokenize it as part
of `2>/dev/null`, when in fact it's part of `>&2`.

To fix the problem, teach wordPart that it can construct a literal even
if it's a _LitRedir, the start of `N>word`. This is fine, because we
already use _LitRedir as a special case where it is needed, and it
should not be a special case here - otherwise the bug appears.

As reported by Sora Morimoto in asdf-community/asdf-direnv#120.