syntax: give a better error when "function" is used in posix mode
Trying to parse a Bash function in posix mode like

	function foo() { bar; }

would result in a rather confusing error:

	1:13: a command can only contain words and redirects; encountered (

It is technically correct, as it parses "function" as a simple command,
and those cannot contain a stray parenthesis.
However, that does not help the user, who might not notice they are
parsing their shell in posix mode rather than bash.

Now the error says:

	1:13: the "function" builtin exists in bash; tried parsing as posix

Do the same for other bash builtins which can fail to parse in a similar
way as simple commands, such as "declare" and "let".

Fixes #911.