expand: avoid panics on division by zero
Arithm now returns one more type of error for this case.
Update the interpreter accordingly, and add tests.

The behavior we adopt follows what Bash does when not run interactively,
so start documenting that in the package godoc.

Also add a TODO as we should have some standard to tell whether errors
coming from the expand package or the user's handlers are fatal.
That is, whether they should exit the shell, and with what status code.

While here, replace black and white with block and allow,
which are more self-explanatory.

Fixes #892.