cmd/shfmt: remove SHFMT_NO_EDITORCONFIG
In late 2019, I wrote:

	We've also added an undocumented SHFMT_NO_EDITORCONFIG environment
	variable, which can be set to "true" to completely disable the feature
	in case it breaks badly when first released. This flag will be removed
	once the feature has been released and stable for a few months.

The rollout of editorconfig support went well enough,
so I don't recall suggesting that any uesrs rely on this fallback.

To my surprise and concern, not only did I forget to remove support for
the environment variable after "a few months",
but quite a few users still use it to this day in 2023.

Remove its support, printing a warning so users are aware.
This time add a TODO as a reminder for proper deletion.