syntax: use fewer newlines in heredocs with minify
Given the shell

  a=$(
    cat <<EOF
    hello
  EOF
  )

We would format it into

  a=$(cat \
  <<EOF)
    hello
  EOF

First, this makes bash unhappy, as it wants to find the closing EOF
before we close the command substitution. Teach rightParen to use a
newline if there are any pending heredocs, even if we're minifying.

Second, the escaped newline before the redirect is unnecessary.

Now, we format the shell into

  a=$(cat <<EOF
    hello
  EOF
  )

Thanks to Andrey Kaipov for the report and test cases.

Fixes #923.