syntax: treat &> as an error in POSIX mode
We were parsing

    foo &>bar

in POSIX mode just as if it were

    foo &
    >bar

That is, a background command, followed by a lone redirect.

This is technically correct per the POSIX shell spec,
since it doesn't have a `&>` token at all,
and the `&` token finishes the command completely.

However, in practice, this behavior is never what the user wants.
If the user feeds this input to the POSIX mode parser,
they are writing Bash syntax but they are using the wrong mode.
This could be if they are writing Bash in a script with extension `.sh`
and no "bash" shebang, for example, as then we assume POSIX mode.
If the file extension is `.bash`, a shebang like `#!/usr/bin/bash`
is used, or the `-ln=bash` flag is used, then the user's code will work.

For that reason, make this edge case a helpful error,
pointing the user towards parsing their code as Bash.

Closes #991, where two users were being confused.