all: assume C.UTF-8 is available in the tests
glibc has shipped with it for well over a year,
and Debian 12 was recently released with a new enough glibc as well.

At this point, anyone testing this Go module on a reasonably recent
Linux system should have the C.UTF-8 locale available.
If someone is running a system without such a locale,
such as an old Linux distro version with an older glibc,
I'm assuming that they are also fine with using an older version
of this repository.