syntax: stop embedding Pos in error struct types
Otherwise the error types gain methods like String or Offset,
which don't make any sense on an error as a whole.
Particularly String, as it won't stringify the entire error;
just its position, omitting the rest of the fields, unlike Error.

This is technically a breaking change, but I never intended the methods
to be included in the error types, and hopefully noone relies on that.
Anyone who relied on them likely got confused by my API mistake here,
and might have buggy code as a result of it.