interp,expand: add support for fs.DirEntry
That is, all of our APIs mirroring the signature of ioutil.ReadDir
now gain a "2" sibling with whose signature mirrors os.ReadDir.
The original APIs are deprecated, and where both can be set together,
the new API takes precedence over the old one.

This should lead to minor overhead reductions by default,
as well as for any users of the old APIs who switch to the new ones.
See the godoc explaining why ioutil.ReadDir,
as well as related APIs like filepath.Walk, are deprecated.