interp: support subshells with FuncEnviron as Env
FuncEnviron can't support listing all env vars via Environ.Each,
and Runner.Subshell relied on that to make a full copy of an Environ.

Instead of trying to make the full copy upfront, which might not work,
the runner now makes a copy of each variable value in setVar
when it is about to be modified, to prevent subshells from modifying
the values in the parent shell.

This method is still not perfect, hence the added TODOs,
but it's certainly more correct as we fix a significant footgun.

Fixes #1043.