syntax: better column numbers for escapes inside backquotes
Backslashes inside backquote command substitutions are deduplicated,
so they are not included in any literal values,
but they still need to be counted for the sake of column numbers.

With the input

    last=`echo \\$demo`

we used to think that the literal with Go string value "\\$demo",
deduplicated from the input "\\\\$demo", went from column 12 to 18.
This can't be, since the input literal was seven bytes and not six,
even if we end up parsing the literal as six bytes after deduplicating.
We now correctly parse the column positions as 12 and 19.

Fixes #1028.