syntax: backquotes affect backslashes in single quotes too
I incorrectly understood that backquote command substitutions
did not affect backslashes in single quotes. For example,

    $ echo `echo "1.2" | sed -e 's/\./\\\\./g'`
    1\.2

would be formatted by shfmt as the following, with different output:

    $ echo $(echo "1.2" | sed -e 's/\./\\\\./g')
    1\\.2

when in fact we do want to deduplicate the backslashes:

    $ echo $(echo "1.2" | sed -e 's/\./\\./g')
    1\.2

Tweak the existing test which expected the wrong result.
Thankfully, we simplify the code now with the fix.

Fixes #1041.