syntax: implement basic offset overflow protection
Since an offset of 0 is valid, we stop increasing the numbers instead.

Also, detect overflows when we construct a position, and not every time
we are advancing the lexer by one character, which caused overhead.
We did this so that a long-lived parser would remember that
once it has overflowed once, all future numbers are also invalid.

However, in practice, using int64 counters for lines and columns
should be more than enough, and saves us the added state.
If anyone has shell input with over 9223372036854775807 lines,
I sure hope that they aren't using my software for it.

            │     old     │                new                 │
            │   sec/op    │   sec/op     vs base               │
    Parse-8   14.86µ ± 1%   14.66µ ± 0%  -1.37% (p=0.000 n=10)