cmd/shfmt: add support for simplify and minify via EditorConfig
This requires treating them as parser and printer flags respectively,
so that they cannot be set as both CLI flags and EditorConfig options
at the same time and lead to confusing results.

While here, regroup parser and printer options so that main.go
and shfmt.1.scd agree on the categorization.

Fixes #819.

Co-authored-by: Daniel Martí <mvdan@mvdan.cc>