syntax: include skipped backslashes in backticks for column numbers
Like a recent commit did for null bytes.
This makes the code simpler to follow, avoiding crashes like the one
that the fuzzer found and is being added here,
and prevents weirdness where p.w does not correspond to len(p.r)
and we need to special case it in Parser.endLit.

This does go a bit in the opposite direction compared to the fix
we did for #1028, where we tried to adjust positions with backslashes
inside backticks. We still count those backslashes for column positions,
but any position of an escaped character points past the backslashes.

Fixes #1098.