syntax: parse all CRLF line endings as LF
We treat CR as whitespace, so this already happened in practice
for regular line endings. However, since heredocs are a quoted state
where white space is not skipped, the CRLF line endings caused errors
where we couldn't match `<<EOF` with a CRLF EOF line as we got `EOF\r`.

Fixes #1088.