fileutil: Omit zsh extensions and shebangs
To prevent `shfmt` from discovering `zsh` files which aren't currently
supported for formatting.

This fixes #535 and the following test:

    > cmpenv stdout find.golden
     diff stdout find.golden
     --- stdout
     +++ find.golden
     @@ -16,6 +16,4 @@
      modify/shebang-space
      modify/shebang-tabs
      modify/shebang-usr-sh
     -skip/ext.zsh
     -skip/shebang-zsh
      symlink/subdir/ext-shebang.sh

     FAIL: testdata/script/walk.txtar:10: stdout and find.golden differ

It could be considered a breaking change to `shfmt` and `fileutil`.

There's some discussion in that issue about whether it's beneficial to
match `zsh` files because they can _sometimes_ be formatted if they
don't contain any zsh-specific syntax, but without any way to explicitly
exclude them from discovery it's more of a hindrance than help.