cmd/shfmt: report correct language read from EditorConfig in errors
We had logic to catch mismatching language errors and report what
language was currently being used for parsing, following the -ln flag.

However, this logic completely ignored EditorConfig,
which can also control this behavior via the shell_variant property.
The logic to determine what language to use for parsing was correct,
but the logic to report a reasonable error to the user was not.

Fixes #1102.