interp: let HandlerContext.Stdin be nil again when there is no stdin
We recently changed Runner.stdin from an io.Reader to an *os.File.
One unintended consequence of this change is that HandlerContext.Stdin
is filled directly from that value, so when the file was nil,
we would now fill Stdin with a typed nil, which is not nil.

This new behavior was unintentional and entirely confusing. Avoid it,
and add a comment to ensure we don't fall into that again.