interp: process substitutions are not waited for
That is, when running

    test -e <(echo foo)

the fact that `echo foo` there is being left hanging should not mean
that `test -e` should be blocked and prevented from finishing.
This is Bash's logic anyway, and doing otherwise breaks scripts.

While here, I also noticed that `wait` is documented to always succeed
when given no arguments, so implement that too.

Fixes #1032.