interp,expand: rethink setting variables
expand.WriteEnviron.Set was a bit of a mess, as it had to support
three very distinct use cases:

1) Changing the value of a variable, such as `foo=bar`

2) Changing an attribute of a variable, such as `export foo`

3) Unsetting a variable, such as `unset foo`

4) Declaring the "kind" of a variable, such as `declare -A foo`

All four had to take into account the existing variable if it was
already declared; for example, `export foo; foo=bar` and
`foo=bar; export foo` should both result in an exported variable
"foo" with string value "bar".

To properly support all the edge cases involved in these,
the expand.Variable API is changed in two ways.

First, whether or not a variable is set is now tracked separately
to its kind. This is because `declare -A foo` has an associative array
kind, but is otherwise still an unset variable. We used to track
unset variables via the zero value "unset" kind, which made this
edge case impossible to track properly.

This includes a new Variable.Declared method, mirroring Variable.IsSet.

Second, to prevent `readonly foo=bar; export foo` into causing an
incorrect "readonly variable" error, the old logic used to have
a hacky heuristic to see if the second command was only trying to add
the "export" attribute without changing its value.
This is now done in a much more explicit way via the KeepValue kind.

While these changes could break users who are directly implementing
or using the WriteEnviron interface, I suspect the number of those
is very close to zero, and the previous interface was very confusing
and could not possibly support some edge cases which needed to work.
Hence, I think the changes above are reasonable to do before a v4.

Fixes #1108.