interp: better support for directories in test -r, -w, -x checks
On Linux, we were considering a directory as read-only given the way
that we implemented the checks via os.OpenFile and os.LookPath.
Implement these checks via unix.Access instead.

On Windows, implement an approximation of unix.Access via os.Lstat,
which seems to do a good enough job for now.

hasPermissionToDir can be folded into this as well.

Fixes #1116.